Add Devise and Angular to the project.
Had to override the sessions controller in devise so that it would use
a different layout - when devise would kick out an auth error, it
normally uses the application layout, which includes the angular set
up and would cause an infinite loop. Also added factory girl to the
project and created a habits controller so there'd be something to
test. Not happy about making such big steps, but it's kind of all of
a piece.

: #devise #angular #db #test