Filter out inactive (ended or not in range) habits.
Added a method to Habit to retreive the active habits, given the end
date of the range. Also made it possible to 'end' a habit, causing it
to disappear from the view. The app now supports pulling date ranges
so we can scroll through old months now. Had to add timestamps to the
habit to make this all work, so changed the original migration that
created the table.

The refactor tag here is because I moved a step from user_steps to
habit_steps.

: #ui #feature #refactor #test